api-app/requirements.txt
# FastAPI and production dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Firebase for cloud storage
firebase-admin>=6.2.0
google-cloud-storage>=2.10.0

# CORS and additional utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP requests
requests>=2.31.0

# Audio processing
torchaudio>=2.1.0
torch>=2.1.0
pydub>=0.25.1

# Text processing
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
